bot/requirements.txt
aiogram
python-dotenv
deep_translator
httpx
requests